pyproject.toml
[project]
name = "apache_airflow_microsoft_fabric_plugin_cs"
version = "1.1.0"
description = "A plugin for Apache Airflow to interact with Microsoft Fabric items"
authors = [
  { name="Ambika Garg", email="ambikagarg1101@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
classifiers = [
    "Programming Language :: Python :: 3",
    "Development Status :: 3 - Alpha",
    "Environment :: Plugins",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
dependencies = [
    "apache-airflow>=2.3.0,<3.0.0",
    "aiohttp>=3.8.0",
    "requests>=2.25.0",
]

[project.urls]
Homepage = "https://github.com/jacobericson/apache-airflow-microsoft-fabric-plugin.git"

[tool.poetry]
packages = [{include = "apache_airflow_microsoft_fabric_plugin", from = "src"}]

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"
pytest-asyncio = ">=0.21.0"
pytest-mock = ">=3.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
